move ResultGetState to obj package
